Simplified the signature of the wrapper function that is generated by 'CFDlang': arguments are now passed as a sequence of pointers.
Previously, arguments were passed as an array of pointers. The new
wrapper signature saves a number of array accesses when the wrapped
function is called inside the wrapper.

The new wrapper signatures require that functions generated by the
'CFDlang' library be called through a function pointer. This, in
turn, saves several calls through the 'CFDlang' library's interface.